# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conferência de Pedidos API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/pedidos": {
      "get": {
        "tags": [
          "pedidos"
        ],
        "summary": "Listar Pedidos",
        "operationId": "listar_pedidos_api_pedidos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Ignora cache e lê de novo no Supabase",
              "default": false,
              "title": "Refresh"
            },
            "description": "Ignora cache e lê de novo no Supabase"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/erros-sheets": {
      "get": {
        "tags": [
          "pedidos"
        ],
        "summary": "Listar Erros Sheets",
        "operationId": "listar_erros_sheets_api_pedidos_erros_sheets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/erros-sheets/count": {
      "get": {
        "tags": [
          "pedidos"
        ],
        "summary": "Contar Erros Sheets",
        "operationId": "contar_erros_sheets_api_pedidos_erros_sheets_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/erros-sheets/resolver": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Resolver Erro Sheets",
        "operationId": "resolver_erro_sheets_api_pedidos_erros_sheets_resolver_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolverSheetsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/erros-sheets/retry": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Retry Erro Sheets",
        "operationId": "retry_erro_sheets_api_pedidos_erros_sheets_retry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolverSheetsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/erros-finalizacao": {
      "get": {
        "tags": [
          "pedidos"
        ],
        "summary": "Listar Erros Finalizacao",
        "operationId": "listar_erros_finalizacao_api_pedidos_erros_finalizacao_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/erros-finalizacao/count": {
      "get": {
        "tags": [
          "pedidos"
        ],
        "summary": "Contar Erros Finalizacao",
        "operationId": "contar_erros_finalizacao_api_pedidos_erros_finalizacao_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/erros-finalizacao/resolver": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Resolver Erro Finalizacao",
        "operationId": "resolver_erro_finalizacao_api_pedidos_erros_finalizacao_resolver_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolverFinalizacaoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/erros-finalizacao/retry": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Retry Erro Finalizacao",
        "operationId": "retry_erro_finalizacao_api_pedidos_erros_finalizacao_retry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolverFinalizacaoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/lote/passo": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Lote Passo",
        "description": "Uma fase de um pedido — o front controla o progresso do modal.",
        "operationId": "lote_passo_api_pedidos_lote_passo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LotePassoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/lote/imprimir": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Imprimir Lote",
        "operationId": "imprimir_lote_api_pedidos_lote_imprimir_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pedidos/despachos/recentes": {
      "get": {
        "tags": [
          "pedidos"
        ],
        "summary": "Listar Despachos Recentes",
        "operationId": "listar_despachos_recentes_api_pedidos_despachos_recentes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 500,
              "title": "Limite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/{order_id}/itens": {
      "get": {
        "tags": [
          "pedidos"
        ],
        "summary": "Itens Do Pedido",
        "operationId": "itens_do_pedido_api_pedidos__order_id__itens_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/{order_id}/presenca": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Heartbeat Presenca",
        "operationId": "heartbeat_presenca_api_pedidos__order_id__presenca_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pedidos"
        ],
        "summary": "Sair Presenca",
        "operationId": "sair_presenca_api_pedidos__order_id__presenca_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/{order_id}/conferir-item": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Conferir Item",
        "operationId": "conferir_item_api_pedidos__order_id__conferir_item_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConferenciaItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/{order_id}/desmarcar-conferencia": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Desmarcar Conferencia",
        "operationId": "desmarcar_conferencia_api_pedidos__order_id__desmarcar_conferencia_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/{order_id}/retomar-finalizacao": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Retomar Finalizacao",
        "operationId": "retomar_finalizacao_api_pedidos__order_id__retomar_finalizacao_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetomarFinalizacaoRequest",
                "default": {
                  "forcar_conferencia": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/{order_id}/imprimir": {
      "post": {
        "tags": [
          "pedidos"
        ],
        "summary": "Imprimir",
        "operationId": "imprimir_api_pedidos__order_id__imprimir_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImprimirRequest",
                "default": {
                  "modo": "todos"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConferenciaItemRequest": {
        "properties": {
          "line_key": {
            "type": "string",
            "minLength": 3,
            "title": "Line Key"
          },
          "quantidade": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantidade"
          },
          "forcar_conferencia": {
            "type": "boolean",
            "title": "Forcar Conferencia",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "line_key",
          "quantidade"
        ],
        "title": "ConferenciaItemRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LotePassoRequest": {
        "properties": {
          "order_id": {
            "type": "string",
            "minLength": 1,
            "title": "Order Id"
          },
          "fase": {
            "type": "string",
            "minLength": 3,
            "title": "Fase",
            "description": "conferir | baixar | imprimir | planilha"
          },
          "forcar_conferencia": {
            "type": "boolean",
            "title": "Forcar Conferencia",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "fase"
        ],
        "title": "LotePassoRequest"
      },
      "LoteRequest": {
        "properties": {
          "order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order Ids"
          }
        },
        "type": "object",
        "required": [
          "order_ids"
        ],
        "title": "LoteRequest"
      },
...
      "ResolverFinalizacaoRequest": {
        "properties": {
          "order_id": {
            "type": "string",
            "minLength": 1,
            "title": "Order Id"
          }
        },
        "type": "object",
        "required": [
          "order_id"
        ],
        "title": "ResolverFinalizacaoRequest"
      },
      "ResolverSheetsRequest": {
        "properties": {
          "order_id": {
            "type": "string",
            "minLength": 1,
            "title": "Order Id"
          },
          "line_key": {
            "type": "string",
            "minLength": 3,
            "title": "Line Key"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "line_key"
        ],
        "title": "ResolverSheetsRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}